requirements.txt
streamlit~=1.10.0
fastf1~=2.2.9
pandas~=1.4.3
plotly~=5.9.0
matplotlib~=3.5.2
numpy>=1.23.0
Pillow>=9.2.0
requests>=2.28.1